image-caption-generator/requirements.txt
flask
werkzeug
torch
torchvision
transformers
opencv-python
pillow
numpy
